Decompose Dialog: extract diff editor header
